+++ fabric/src/main/java/io/github/sakurawald/module/mixin/chat/style/message_type/RegistryLoaderMixin.java
@@ -48,6 +48,7 @@ import java.util.Map;
 public class RegistryLoaderMixin {
 
     #if MC_VER <= MC_1_20_4
+    @SuppressWarnings("LocalCaptureFailException")
     @Inject(method = "load(Lnet/minecraft/resource/ResourceManager;Lnet/minecraft/registry/DynamicRegistryManager;Ljava/util/List;)Lnet/minecraft/registry/DynamicRegistryManager$Immutable;"
         , at = @At(value = "INVOKE", target = "Ljava/util/List;forEach(Ljava/util/function/Consumer;)V", ordinal = 0, shift = At.Shift.AFTER), locals = LocalCapture.CAPTURE_FAILEXCEPTION)
     private static void registerNewMessageType(
@@ -86,12 +87,12 @@ public class RegistryLoaderMixin {
             MutableRegistry<?> registry = entry.comp_2246();
             #endif
 
+            /* Register out custom message type in the proper registry. */
             RegistryKey<? extends Registry<?>> registryKey = registry.getKey();
-
             if (registryKey.equals(RegistryKeys.MESSAGE_TYPE)) {
                 Registry<MessageType> registryForMessageType = (Registry<MessageType>) registry;
 
-                // note: in single-player world, the MESSAGE_TYPE_KEY will be registered twice, causing a `network protocol error` while join the world.
+                // NOTE: in single-player world, the MESSAGE_TYPE_KEY will be registered twice, causing a `network protocol error` while join the world.
                 if (!registryForMessageType.contains(ChatStyleInitializer.MESSAGE_TYPE_KEY)) {
                     Registry.register(registryForMessageType, ChatStyleInitializer.MESSAGE_TYPE_KEY, ChatStyleInitializer.MESSAGE_TYPE_VALUE);
                 }